Perbaiki Kurir

+++ src/main/java/com/recycleme/actionListener/kurir/KurirHapus.java
@@ -4,6 +4,7 @@ import com.recycleme.dao.KurirDao;
 import com.recycleme.frame.kurir.KurirFrame;
 import com.recycleme.model.kurir.Kurir;
 
+import javax.swing.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
@@ -20,12 +21,27 @@ public class KurirHapus implements ActionListener {
 
     public void actionPerformed(ActionEvent e) {
         if(e.getSource() == kurirFrame.getButtonDeleteKurir()) {
+            int row = kurirFrame.getSelectedKurirRow();
+            if(row == -1) {
+                kurirFrame.showErrorMessage("Pilih kurir terlebih dahulu!");
+                return;
+            }
+
             int id = kurirFrame.getSelectedKurirId();
+
+            int confirm = JOptionPane.showConfirmDialog(kurirFrame, "Apakah anda yakin ingin menghapus masyarakat ini?",
+                    "Konfirmasi", JOptionPane.YES_NO_OPTION);
+
+            if(confirm == JOptionPane.NO_OPTION) {
+                kurirFrame.showInfoMessage("Proses dibatalkan!");
+                return;
+            }
+
             if(kurirDao.delete(id) == 0) {
                 kurirFrame.showErrorMessage("Data gagal dihapus!");
                 return;
             }
-            kurirFrame.removeKurir(id);
+            kurirFrame.removeKurir(row);
             kurirFrame.showSuccessMessage("Data berhasil dihapus!");
         }
     }
